rename : tag to colType
