+++ plugindata.json
@@ -1,150 +1,164 @@
 {
-  "version": "0.14.0",
+  "version": "0.15.0",
   "data": {
     "boardConfigs": [
       {
         "columns": [
           {
-            "tag": "undated",
+            "colType": "undated",
             "data": {
-              "collapsed": false,
-              "name": "Undated"
-            }
+				"collapsed": false,
+				"name": "Undated",
+				"index": 1
+			}
           },
           {
-            "tag": "dated",
+            "colType": "dated",
             "data": {
-              "collapsed": false,
-              "name": "Over Due",
-              "range": {
-                "tag": "before",
-                "rangedata": {
-                  "from": -1,
-                  "to": 0
-                }
-              }
-            }
+				"collapsed": false,
+				"name": "Over Due",
+				"index": 2,
+				"range": {
+					"tag": "before",
+					"rangedata": {
+						"from": -1,
+						"to": 0
+					}
+				}
+			}
           },
           {
-            "tag": "dated",
+            "colType": "dated",
             "data": {
-              "collapsed": false,
-              "name": "Today",
-              "range": {
-                "tag": "between",
-                "rangedata": {
-                  "from": 0,
-                  "to": 0
-                }
-              }
-            }
+				"collapsed": false,
+				"name": "Today",
+				"index": 3,
+				"range": {
+					"tag": "between",
+					"rangedata": {
+						"from": 0,
+						"to": 0
+					}
+				}
+			}
           },
           {
-            "tag": "dated",
+            "colType": "dated",
             "data": {
-              "collapsed": false,
-              "name": "Tomorrow",
-              "range": {
-                "tag": "between",
-                "rangedata": {
-                  "from": 1,
-                  "to": 1
-                }
-              }
-            }
+				"collapsed": false,
+				"name": "Tomorrow",
+				"index": 4,
+				"range": {
+					"tag": "between",
+					"rangedata": {
+						"from": 1,
+						"to": 1
+					}
+				}
+			}
           },
           {
-            "tag": "dated",
+            "colType": "dated",
             "data": {
-              "collapsed": false,
-              "name": "Future",
-              "range": {
-                "tag": "after",
-                "rangedata": {
-                  "from": 2,
-                  "to": 0
-                }
-              }
-            }
+				"collapsed": false,
+				"name": "Future",
+				"index": 5,
+				"range": {
+					"tag": "after",
+					"rangedata": {
+						"from": 2,
+						"to": 0
+					}
+				}
+			}
           },
           {
-            "tag": "completed",
+            "colType": "completed",
             "data": {
               "collapsed": false,
-              "index": 5,
               "limit": 10,
-              "name": "Completed"
+              "name": "Completed",
+              "index": 6
             }
           },
           {
-            "tag": "custom",
+            "colType": "custom",
             "data": {
-              "collapsed": false,
-              "name": "Column 7"
-            }
+				"collapsed": false,
+				"name": "Column 7",
+				"index": 7
+			}
           }
         ],
         "filters": [],
         "filterPolarity": "Allow",
         "filterScope": "Both",
         "name": "Pending Tasks",
+		"index": 1,
         "showColumnTags": true,
         "showFilteredTags": true
       },
       {
         "columns": [
           {
-            "tag": "untagged",
+            "colType": "untagged",
             "data": {
-              "collapsed": false,
-              "name": "Untagged"
-            }
+				"collapsed": false,
+				"name": "Untagged",
+				"index": 1
+			}
           },
           {
-            "tag": "namedTag",
+            "colType": "namedTag",
             "data": {
-              "collapsed": false,
-              "name": "GPT",
-              "coltag": "#OfflineVideo"
-            }
+				"collapsed": false,
+				"name": "GPT",
+				"index": 2,
+				"coltag": "#OfflineVideo"
+			}
           },
           {
-            "tag": "namedTag",
+            "colType": "namedTag",
             "data": {
-              "collapsed": false,
-              "name": "When Mobile Data Left at Night",
-              "coltag": "#WhenMobileDataLeft"
-            }
+				"collapsed": false,
+				"name": "When Mobile Data Left at Night",
+				"index": 3,
+				"coltag": "#WhenMobileDataLeft"
+			}
           },
           {
-            "tag": "namedTag",
+            "colType": "namedTag",
             "data": {
-              "collapsed": false,
-              "name": "Note Till Task is Done",
-              "coltag": "#NOTE"
-            }
+				"collapsed": false,
+				"name": "Note Till Task is Done",
+				"index": 4,
+				"coltag": "#NOTE"
+			}
           },
           {
-            "tag": "otherTags",
+            "colType": "otherTags",
             "data": {
-              "collapsed": false,
-              "name": "Other Tags"
-            }
+				"collapsed": false,
+				"name": "Other Tags",
+				"index": 5
+			}
           },
           {
-            "tag": "completed",
+            "colType": "completed",
             "data": {
-              "collapsed": false,
-              "index": 5,
-              "limit": 10,
-              "name": "Completed"
-            }
+				"collapsed": false,
+				"index": 5,
+				"limit": 10,
+				"name": "Completed",
+				"index": 6
+			}
           }
         ],
         "filters": [],
         "filterPolarity": "Allow",
         "filterScope": "Both",
         "name": "Tag Based Board",
+		"index": 2,
         "showColumnTags": true,
         "showFilteredTags": true
       }
