+++ tasks.json
@@ -92,27 +92,23 @@
         "tag": "",
         "filePath": "Spintly Details.md",
         "completed": false
-      },
-      {
-        "id": 1725886490784,
-        "body": "New tasks testing",
-        "due": "2024-09-09",
-        "tag": "#test"
       }
     ],
     "Daily_Notes/07-09-2024.md": [
-      {
-        "id": 1725690615381,
-        "body": "ChatGPT sometimes do things amazing",
-        "due": "2024-09-08",
-        "tag": "#GPT"
-      },
       {
         "id": 1725697862884,
         "body": "Just adding a small task to get GPT tag.\nJust adding a small task to get GPT tag.\nJust adding a small task to get GPT tag.",
         "due": "2024-09-12",
         "tag": "#GPT",
         "filePath": "Daily_Notes/07-09-2024.md"
+      },
+      {
+        "id": 1725690615381,
+        "body": "ChatGPT sometimes do things amazing",
+        "due": "2024-09-08",
+        "tag": "#GPT",
+        "filePath": "Daily_Notes/07-09-2024.md",
+        "completed": false
       }
     ],
     "Daily_Notes/08-09-2024.md": [],
@@ -170,7 +166,16 @@
     ],
     "Daily_Notes/03-09-2024.md": [],
     "Daily_Notes.md": [],
-    "Spintly Details.md": [],
+    "Spintly Details.md": [
+      {
+        "id": 1725886490784,
+        "body": "New tasks testing",
+        "due": "2024-09-09",
+        "tag": "#test",
+        "filePath": "Spintly Details.md",
+        "completed": true
+      }
+    ],
     "Daily_Notes/08-09-2024.md": [
       {
         "id": 1725785322765,
@@ -180,6 +185,7 @@
         "filePath": "Daily_Notes/08-09-2024.md",
         "completed": true
       }
-    ]
+    ],
+    "Daily_Notes/07-09-2024.md": []
   }
 }
